App разложен по логическим кусочкам
