+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 dist
 example
 notice.txt
+order.txt
