+++ src/app.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useMemo, forwardRef } from "react";
 import ReactDOM from "react-dom";
 import { Provider, useDispatch, useSelector } from "react-redux";
-import { createStore, compose } from "redux";
 import styled, { ThemeProvider } from "styled-components";
 
 import Grid from "./features/Grid";
@@ -10,15 +9,11 @@ import Dice from "./features/Dice";
 import StartScreen from "./features/StartScreen";
 import EndScreen from "./features/EndScreen";
 
-import waitingStartPhase from "./phases/waitingStart";
-import trownDice from "./phases/gameStarted/trownDice";
-import clickArrow from "./phases/gameStarted/clickArrow";
-import takeHealthCard, {
-  getManHealth,
-} from "./phases/gameStarted/takeHealthCard";
-import getOrder from "./phases/gameStarted/getOrder";
-import endGame from "./phases/endGame";
-
+import endGame from "./business/phases/endGame";
+import { reducer } from "./business/reducer";
+import { State } from "./business/types";
+import { GameList, CoordItem } from "./business/types";
+import { store } from "./business/store";
 const Field = styled.div`
   position: relative;
   width: 300px;
@@ -47,539 +42,7 @@ const Status = styled.div`
   min-height: 18px;
 `;
 
-/** параметризируемые переменные  **/
 
-export const StartCell = { hor: 0, vert: 0 };
-export const EndCell = { hor: 9, vert: 9 };
-const initialManHealth = 3;
-const amountHealthItems = 30;
-export const amountMen = 4;
-/*зачем minNumerbMan -?! */
-export const minNumerbMan = 1;
-
-const wallList: Array<CoordItem> = [
-  { hor: 2, vert: 2 },
-  { hor: 3, vert: 2 },
-  { hor: 4, vert: 2 },
-  { hor: 2, vert: 3 },
-  { hor: 4, vert: 3 },
-  { hor: 2, vert: 4 },
-  { hor: 3, vert: 4 },
-  { hor: 4, vert: 4 },
-];
-
-/** объявление типов  **/
-export type CoordItem = { hor: number; vert: number };
-
-export type HealthItemType = "increment" | "decrement";
-
-export type HealthItemTypeArr = ["increment", "decrement"];
-
-export type MoveDirection = "top" | "bottom" | "left" | "right";
-
-export type WallItem = {
-  name: "wall";
-};
-
-export type ManItem = {
-  name: "man";
-  health: number;
-  orderNumber: number;
-};
-
-export type ManList = ManItem[];
-
-export type FinishCell = {
-  name: "finish";
-  cardItem: { manList?: ManList };
-};
-export type StartCell = {
-  name: "start";
-  cardItem: { manList?: ManList };
-};
-
-export type HealthItem = {
-  name: "health";
-  type: HealthItemType;
-  apperance: "closed" | "open";
-};
-
-// почему ManList а не HealthItem ?! потому что несколько человек может на 1 клетке стоять?!
-export type CardInteract = ManList | HealthItem;
-
-/*нужен для рисования массива здоровья */
-export type FieldItem = {
-  hor: number;
-  vert: number;
-  name: "field";
-  cardItem: CardInteract[];
-};
-
-export type ObjHealthItem = {
-  hor: number;
-  vert: number;
-  name: "field";
-  cardItem: { healthItem: HealthItem };
-};
-
-export type ManAndHealthFieldItem = {
-  name: "field";
-  cardItem: { manList: ManList; healthItem: HealthItem };
-};
-export type EmptyFieldItem = {
-  name: "field";
-};
-export type ObjFieldItem = {
-  name: "field";
-  /*   cardItem: { manItem?: ManItem; healthItem?: HealthItem }; */
-  cardItem: { manList?: ManList; healthItem?: HealthItem };
-};
-export type ObjCellType = ObjFieldItem | WallItem | FinishCell | StartCell;
-
-export type GameList = Map<string, ObjCellType>;
-
-export type TypeEffect =
-  | { type: "!needOpenHealthCard" }
-  | { type: "!changeManHealth" }
-  | { type: "!changeHealthList" }
-  | { type: "!getNextMan" }
-  | null;
-
-export type State = {
-  gameState: GameState;
-  dice: number;
-  gameResult: "" | "Вы выиграли" | "Вы проиграли";
-  cardInteractIndex: string[];
-  GameList: GameList;
-  doEffect: TypeEffect;
-  numberOfMan: number;
-};
-
-export type ActionType =
-  | { type: "clickStartButton" }
-  | DiceThrownAction
-  | ArrowPressAction
-  | { type: "openedHealthCard" }
-  | { type: "changedManHealth" }
-  | { type: "changedHealthList" }
-  | { type: "receivedNextMan" }
-  | { type: "getEndScreen" };
-
-export type ArrowPressAction = { type: "arrowPressed"; payload: MoveDirection };
-export type DiceThrownAction = { type: "diceThrown"; payload: number };
-
-export type GameState =
-  | { type: "waitingStart" /* context: {}  */ }
-  | { type: "gameStarted.trownDice"; context: any; gameStartedContext: any }
-  | {
-      type: "gameStarted.clickArrow";
-      gameStartedContext: any;
-      context: any;
-    }
-  | openHealthCardType
-  | { type: "gameStarted.getOrder" }
-  | { type: "endGame"; context: any }
-  | { type: "getEndScreen"; context: any };
-
-export type openHealthCardType = {
-  type: "gameStarted.takeHealthCard";
-  /*   gameStartedContext: any; */
-  context: any;
-};
-
-//зачем нужен дополнительный к HealthItemTypeArr тип-?!
-export const healthItemTypeArr: HealthItemTypeArr = ["increment", "decrement"];
-//____________________2 вариант 120 строк
-//сначала ставим стены, потом рандомно раскладываем карточки здоровья
-
-
-//1.создаем пустую Map-у со стенами, стартом и финишем
-
-
-const createGameField = (number: number, endCell: CoordItem): any => {
-  const width = EndCell.hor;
-  const height = EndCell.vert;
-
-  const emptyFieldItem: ObjFieldItem = {
-    name: "field",
-    cardItem: {},
-  };
-
-  const finishCell: FinishCell = {
-    name: "finish",
-    cardItem: {},
-  };
-  const startCell: StartCell = {
-    name: "start",
-    cardItem: {},
-  };
-
-  const emptyMap = new Map();
-
-  for (let hor = 0; hor <= width; hor++) {
-    for (let vert = 0; vert <= height; vert++) {
-      const hasFinish = width === hor && height === vert;
-      const hasStart = hor === 0 && vert === 0;
-      const wallCell = wallList.find((item) => {
-        return item.hor === hor && item.vert === vert;
-      });
-      const hasWall = wallCell ? true : false;
-      switch (true) {
-        case hasWall: {
-          const wallItem: WallItem = {
-            name: "wall",
-          };
-          emptyMap.set(`${hor}.${vert}`, wallItem);
-          break;
-        }
-        case hasFinish: {
-          emptyMap.set(`${hor}.${vert}`, finishCell);
-          break;
-        }
-        case hasStart: {
-          emptyMap.set(`${hor}.${vert}`, startCell);
-          break;
-        }
-        default:
-          emptyMap.set(`${hor}.${vert}`, emptyFieldItem);
-      }
-    }
-  }
-  return emptyMap;
-};
-
-//2.раскладываем карточки здоровья
-//изначально карточки лежат по одной на ячейке
-
-const spreadCards = (initialField: GameList) => {
-  
-  const arrayGameField: [string, ObjCellType][] = Array.from(initialField);
-
-  //emptyItems - пустые карточки
-  const emptyItems: [string, ObjCellType][] = arrayGameField.filter(
-    (innerArr) => {
-      return innerArr.find((item) => {
-        //только один объект в ячейке
-        return item instanceof Object && item.name === "field";
-      });
-    }
-  );
-
-  let maxRandomNumber = emptyItems.length;
-
-  const indexArray: Number[] = takeItemForCard(
-    amountHealthItems,
-    maxRandomNumber
-  );
-
-  function getRandomType(): HealthItemType {
-    return healthItemTypeArr[Math.floor(Math.random() * 2)];
-  }
-
-  let itemsWithHealth = emptyItems.map((item, index) => {
-    let cardWithHealth = indexArray.find((indexItem) => indexItem === index);
-    //по этому приципу можно будет раскладывать и карточки противников!
-
-    if (cardWithHealth) {
-      //положить по этим ячейкам карточки здоровья
-      const [index, value] = item;
-      if (value.name === "field") {
-        const healthCell: ObjCellType = {
-          ...value,
-          cardItem: {
-            ...value.cardItem,
-            healthItem: {
-              name: "health",
-              type: getRandomType(),
-              apperance: "closed",
-            },
-          },
-        };
-        return [index, healthCell] as [string, ObjCellType];
-      } else return item;
-    } else return item;
-  });
-
-  let fullGameField = [...arrayGameField, ...itemsWithHealth];
-  let fullMapGameField = new Map(fullGameField);
-
-  return fullMapGameField;
-
-  function takeItemForCard(lenght: number, maxNumber: number): Number[] {
-    //массив из 30 неповторяющихся индексов карточек
-    return new Array(lenght).fill(0).reduce((prev, item) => {
-      const number = getRandomNumber(prev);
-      if (prev) {
-        return [number, ...prev];
-      } else return [number];
-
-      function getRandomNumber(arr: Number[]): Number {
-        let number = Math.floor(Math.random() * maxNumber);
-        const repeat: Boolean =
-          prev && prev.find((item: Number) => item === number) >= 0;
-        if (repeat) {
-          return getRandomNumber(prev);
-        } else return number;
-      }
-    });
-  }
-};
-
-const addMens = (fieild: GameList) => {
-  fieild.forEach((item, index) => {
-    /* const [index, value] = item; */
-    if (item.name === "start") {
-      const startCard: ObjCellType = {
-        name: "field",
-        cardItem: {
-          manList: new Array(amountMen).fill(0).map((item, index) => {
-            return {
-              name: "man",
-              health: initialManHealth,
-              orderNumber: index,
-            };
-          }),
-        },
-      };
-      fieild.set(index, startCard);
-    } else return item;
-  });
-  console.log(fieild);
-  return fieild;
-};
-
-const getNewGameList = (
-  amountHealthItems: number,
-  wallList: Array<CoordItem>,
-  endCell: CoordItem
-) => {
-  const emptyGameField: GameList = createGameField(amountHealthItems, EndCell);
-  const filledCardsField: GameList = spreadCards(emptyGameField);
-  const fullPreparedField = addMens(filledCardsField);
-  return fullPreparedField;
-};
-
-/* console.log("spreadCards", spreadCards()); */
-//____________________1 вариант 168 строк и куча повторений в проверках на стену..
-const getRandomHealthItem = (arr: Array<FieldItem>): ObjHealthItem => {
-  const hor = Math.floor(Math.random() * (EndCell.hor + 1));
-  const vert = Math.floor(Math.random() * (EndCell.vert + 1));
-
-  const randomType: HealthItemType =
-    healthItemTypeArr[Math.floor(Math.random() * 2)];
-
-  const crossStartCell = StartCell.hor === hor && StartCell.vert === vert;
-  const wallCell = wallList.find((item) => {
-    return item && item.hor === hor && item.vert === vert;
-  });
-
-  const crossWall = wallCell ? true : false;
-  const hasIntersection = crossWall || crossStartCell;
-  const healthRepetition = arr.find((item) => {
-    return item && item.hor === hor && item.vert === vert;
-  });
-
-  const hasHealthRepitition = healthRepetition
-    ? healthRepetition.hor === hor && healthRepetition.vert === vert
-    : false;
-  const firstElement = arr.length === 0;
-
-  switch (true) {
-    case hasIntersection: {
-      return getRandomHealthItem(arr);
-    }
-    case !crossStartCell: {
-      switch (true) {
-        case firstElement: {
-          return {
-            hor: hor,
-            vert: vert,
-            name: "field",
-            cardItem: {
-              healthItem: {
-                name: "health",
-                type: randomType,
-                apperance: "closed",
-              },
-            },
-          };
-        }
-        case !hasHealthRepitition: {
-          return {
-            hor: hor,
-            vert: vert,
-            name: "field",
-            cardItem: {
-              healthItem: {
-                name: "health",
-                type: randomType,
-                apperance: "closed",
-              },
-            },
-          };
-        }
-        case hasHealthRepitition: {
-          return getRandomHealthItem(arr);
-        }
-      }
-    }
-    default:
-      return getRandomHealthItem(arr);
-  }
-};
-
-const createHealthArray = (number: number): ObjHealthItem[] => {
-  let healthArray = new Array(number).fill(0).reduce((prevValue) => {
-    const currValue = getRandomHealthItem(prevValue);
-
-    return [currValue, ...prevValue];
-  }, []);
-  console.log(healthArray);
-  return healthArray;
-};
-
-const getGameList = (
-  amountHealthItems: number,
-  wallList: Array<CoordItem>,
-  endCell: CoordItem
-) => {
-  const width = EndCell.hor;
-  const height = EndCell.vert;
-  const healthList: Array<ObjHealthItem> = createHealthArray(amountHealthItems);
-  const gameArray: GameList = new Map();
-
-  for (let hor = 0; hor <= width; hor++) {
-    for (let vert = 0; vert <= height; vert++) {
-      const hasManInStart = StartCell.hor === hor && StartCell.vert === vert;
-      const hasFinish = EndCell.hor === hor && EndCell.vert === vert;
-
-      const health = healthList.find((item, index) => {
-        /*сделать поиск по индексу?!*/
-        return item.hor === hor && item.vert === vert;
-      });
-
-      const hasHealth = health ? true : false;
-      const hasManAndHealth = hasHealth && hasManInStart;
-
-      const wallCell = wallList.find((item) => {
-        return item.hor === hor && item.vert === vert;
-      });
-      const hasWall = wallCell ? true : false;
-
-      const emptyFieldItem: ObjFieldItem = {
-        name: "field",
-        cardItem: {},
-      };
-
-      const getCell = () => {
-        switch (true) {
-          case hasWall: {
-            const wallItem: WallItem = {
-              name: "wall",
-            };
-            return wallItem;
-          }
-          case !hasWall: {
-            switch (true) {
-              case hasFinish: {
-                const finishCell: FinishCell = {
-                  name: "finish",
-                  cardItem: {},
-                };
-                return finishCell;
-              }
-              case hasHealth: {
-                if (health != undefined) {
-                  return health;
-                } else return emptyFieldItem;
-              }
-              case hasManInStart: {
-                const fieldItem: ObjFieldItem = {
-                  name: "field",
-                  //пока грубо поставим 2 человечков
-                  cardItem: {
-                    manList: new Array(amountMen).fill(0).map((item, index) => {
-                      return {
-                        name: "man",
-                        health: initialManHealth,
-                        orderNumber: index,
-                      };
-                    }),
-                  },
-                  /* cardItem: {
-                    manItem: { name: "man", health: initialManHealth },
-                  }, */
-                };
-                return fieldItem;
-              }
-
-              default:
-                return emptyFieldItem;
-            }
-          }
-          default:
-            return emptyFieldItem;
-        }
-      };
-
-      gameArray.set(`${hor}.${vert}`, getCell());
-    }
-  }
-  console.log(gameArray);
-  return gameArray;
-};
-
-const getInitialState = (): State => {
-  return {
-    gameState: { type: "waitingStart" },
-    dice: 0,
-    gameResult: "",
-    cardInteractIndex: new Array(amountMen).fill(0).map((item, index) => {
-      return `${StartCell.hor}.${StartCell.vert}`;
-    }),
-    GameList: getNewGameList(amountHealthItems, wallList, EndCell),
-    doEffect: null,
-    numberOfMan: 0,
-  };
-};
-
-const reducer = (state = getInitialState(), action: ActionType): State => {
-  const [phaseOuter, phaseInner] = state.gameState.type.split(".");
-
-  switch (phaseOuter) {
-    case "waitingStart": {
-      return waitingStartPhase(action, state);
-    }
-
-    case "gameStarted": {
-      switch (phaseInner) {
-        case "trownDice": {
-          return trownDice(action, state);
-        }
-
-        case "clickArrow": {
-          return clickArrow(action, state);
-        }
-
-        case "takeHealthCard": {
-          const gameState = state.gameState as openHealthCardType;
-          return takeHealthCard(action, state, gameState);
-        }
-        case "getOrder": {
-          return getOrder(action, state);
-        }
-        default:
-          return state;
-      }
-    }
-
-    case "endGame": {
-      endGame(action, state);
-    }
-    default:
-      return state;
-  }
-};
 
 const showManListHealth = (
   gameList: GameList,
@@ -595,7 +58,6 @@ const showManListHealth = (
       return manElem ? manElem?.health : null;
     } else return null;
   });
-  console.log(healthArray);
 
   return healthArray;
 };
@@ -610,6 +72,7 @@ function App() {
   } = useSelector((state: State) => ({ ...state }));
 
   const dispatch = useDispatch();
+
   const textPhase = () => {
     switch (gameState.type) {
       case "gameStarted.trownDice":
@@ -722,12 +185,6 @@ function App() {
 
   return <Game>{getGameScreen()}</Game>;
 }
-
-const store = createStore(
-  reducer,
-  (window as any).__REDUX_DEVTOOLS_EXTENSION__ &&
-    (window as any).__REDUX_DEVTOOLS_EXTENSION__()
-);
 
 ReactDOM.render(
   <Provider store={store}>
