+++ src/features/Dice.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 
-import {State} from "./../app"
+import { State } from "../business/types";
 const DiceWrap = styled.div`
   border: 1px solid lightgray;
   width: 100px;
@@ -12,7 +12,6 @@ const DiceWrap = styled.div`
 
 const DiceButton = styled.button`
   margin: 0 auto;
-
 `;
 const DiceNumber = styled.div`
   height: 30px;
@@ -20,14 +19,11 @@ const DiceNumber = styled.div`
   text-align: center;
 `;
 
-
 const maxNumber = 4;
 
 function Dice() {
   const dispath = useDispatch();
-  const [diceNumber] = useSelector((state:State) => [
-    state.dice
-  ]);
+  const [diceNumber] = useSelector((state: State) => [state.dice]);
   const getNumber = () => {
     return Math.floor(Math.random() * maxNumber + 1);
   };
