+++ src/features/Health.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import styled from "styled-components";
 
-import { HealthItem } from "./../app";
+import { HealthItem } from "../business/types";
 
 const HealthItem = styled.div<HealthItem>`
   border: 5px solid;
